refactor: historyWindowをtsに移行
